pyproject.toml
[tool.poetry]
name = "song_translator_flask"
version = "0.1.0"
description = ""
authors = ["Anatoli"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
SQLAlchemy = "^1.4.19"
Flask-RESTful = "^0.3.9"
flask-marshmallow = "^0.14.0"
psycopg2 = "^2.9.1"
python-dotenv = "^0.18.0"
Flask-Migrate = "^3.0.1"
marshmallow-sqlalchemy = "^0.26.1"
google-trans-new = "^1.1.9"
requests = "^2.25.1"
uuid = "^1.30"
PyJWT = "^2.1.0"
DateTime = "^4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
